fix #17
